fix(ir): case: div by (neg) one
and also removed some temp files
